Fix warnings for Win32 Compiler.cpp
This also involves regularising type for GlyphIDs to gid16 instead of
the mishmash of int and utf16 types currently used.
Set version fields to be uint32_t consisntently.

+++ compiler/Grammar/TokenBuffer.cpp
@@ -61,7 +61,7 @@ void TokenBuffer::fill(int amount)
 {
 	syncConsume();
 	// Fill the buffer sufficiently to hold needed tokens
-	while (queue.entries() < amount + markerOffset) {
+	while (queue.entries() < size_t(amount + markerOffset)) {
 		// Append the next token
 		queue.append(input.nextToken());
 	}
